clocksource: dw_apb_timer: add quirk for variants with two value registers
This adds a quirk for IP variants containing two load_count and value
registers that are used to provide for example 64bit accuracy on 32bit
systems.
The added accuracy is currently not used, the driver is only adapted to
handle the different register layout and make it work on affected devices.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>